+++ R/rules.R
@@ -24,42 +24,6 @@ to_output <- function(lf, output, cmd_options = c("--standalone", "--self-contai
     output
 }
 
-#' Get names of piecepack games we can generate rulesets for.
-#'
-#' `names_rulesets()` returns the names of piecepack games we can generate rulesets for.
-#' @param book Book name or `NULL` (for all supported rules).
-#'             Currently only supports "the historical piecepacker".
-#' @seealso [save_ruleset()], [save_pamphlet()], and [save_pocketmod()].
-#' @export
-names_rulesets <- function(book = NULL) {
-    if (is.null(book)) {
-        names <- list.files(system.file("rules", package = "ppgames"))
-        names <- grep(".Rtex$", names, value=TRUE)
-        names <- gsub(".Rtex$", "", names)
-        names <- names[-grep("alice|seasons|ultima", names)]
-        names <- gsub("mens-morris", "men's-morris", names)
-        names <- gsub("-", " ", names)
-    } else {
-        book <- normalize_name(book, sep = "-")
-        stopifnot(book == "the-historical-piecepacker")
-        names <- c("alquerque",
-                   "american checkers",
-                   "awithlaknannai mosona",
-                   "backgammon",
-                   "chaturaji",
-                   "cribbage",
-                   "four field kono",
-                   "international chess",
-                   "ludo",
-                   "nine men's morris",
-                   "tablut",
-                   "twelve men's morris",
-                   "xiangqi")
-    }
-    names
-}
-
-
 # nolint start
 # to_pdf <- function(kf) {
 #     pdf <- sub(paste0(file_ext(kf), "$"), "pdf", kf)
@@ -160,7 +124,7 @@ save_ruleset <- function(game, gk = game_kit(), output = NULL,
 }
 
 setup_tempdir <- function(output) {
-    dir <- file.path(tempdir(), basename(output))
+    dir <- file.path(tempdir(), paste0(basename(output), "_tempdir"))
     unlink(dir, recursive = TRUE)
     dir.create(dir)
     dir